refactor: address review comments — move settle helpers to service layer; fix PAUSED ordering bug
- Move `latest_signal_ts` and `compute_effective_settle` out of
  `models/review.py` into a new `services/review_settle.py` module
  so the models layer stays data-only (Pydantic BaseModel subclasses)
- Fix ordering bug: PAUSED check now runs before the `age >= settle`
  early return, so a paused bot always yields the full settle wait
  regardless of signal age
- Update all call sites (review_service.py, test_review.py) to import
  from the new module
- Add test: paused bot with age >= settle still returns full settle
